+++ fresh.gen.ts
@@ -13,18 +13,18 @@ import * as $6 from "./routes/certificates.tsx";
 import * as $7 from "./routes/index.tsx";
 import * as $8 from "./routes/me.tsx";
 import * as $9 from "./routes/projects/expenses-tracker.tsx";
-import * as $10 from "./routes/projects/food-order.tsx";
-import * as $11 from "./routes/projects/index.tsx";
-import * as $12 from "./routes/tools/highlighted-text/[text].tsx";
-import * as $13 from "./routes/tools/index.tsx";
-import * as $14 from "./routes/tools/retirement-calculator.tsx";
-import * as $15 from "./routes/tools/syntax-highlight.tsx";
-import * as $16 from "./routes/toys/index.tsx";
-import * as $17 from "./routes/toys/insanity.tsx";
-import * as $18 from "./routes/toys/spinners.tsx";
-import * as $19 from "./routes/what-is-this.tsx";
-import * as $20 from "./routes/work/form.tsx";
-import * as $21 from "./routes/work/index.tsx";
+import * as $11 from "./routes/projects/food-order/index.tsx";
+import * as $12 from "./routes/projects/index.tsx";
+import * as $13 from "./routes/tools/highlighted-text/[text].tsx";
+import * as $14 from "./routes/tools/index.tsx";
+import * as $15 from "./routes/tools/retirement-calculator.tsx";
+import * as $16 from "./routes/tools/syntax-highlight.tsx";
+import * as $17 from "./routes/toys/index.tsx";
+import * as $18 from "./routes/toys/insanity.tsx";
+import * as $19 from "./routes/toys/spinners.tsx";
+import * as $20 from "./routes/what-is-this.tsx";
+import * as $21 from "./routes/work/form.tsx";
+import * as $22 from "./routes/work/index.tsx";
 import * as $$0 from "./islands/AddNewExpense.tsx";
 import * as $$1 from "./islands/Collapsible.tsx";
 import * as $$2 from "./islands/CopyTextAreaToClipboard.tsx";
@@ -51,18 +51,18 @@ const manifest = {
     "./routes/index.tsx": $7,
     "./routes/me.tsx": $8,
     "./routes/projects/expenses-tracker.tsx": $9,
-    "./routes/projects/food-order.tsx": $10,
-    "./routes/projects/index.tsx": $11,
-    "./routes/tools/highlighted-text/[text].tsx": $12,
-    "./routes/tools/index.tsx": $13,
-    "./routes/tools/retirement-calculator.tsx": $14,
-    "./routes/tools/syntax-highlight.tsx": $15,
-    "./routes/toys/index.tsx": $16,
-    "./routes/toys/insanity.tsx": $17,
-    "./routes/toys/spinners.tsx": $18,
-    "./routes/what-is-this.tsx": $19,
-    "./routes/work/form.tsx": $20,
-    "./routes/work/index.tsx": $21,
+    "./routes/projects/food-order/index.tsx": $11,
+    "./routes/projects/index.tsx": $12,
+    "./routes/tools/highlighted-text/[text].tsx": $13,
+    "./routes/tools/index.tsx": $14,
+    "./routes/tools/retirement-calculator.tsx": $15,
+    "./routes/tools/syntax-highlight.tsx": $16,
+    "./routes/toys/index.tsx": $17,
+    "./routes/toys/insanity.tsx": $18,
+    "./routes/toys/spinners.tsx": $19,
+    "./routes/what-is-this.tsx": $20,
+    "./routes/work/form.tsx": $21,
+    "./routes/work/index.tsx": $22,
   },
   islands: {
     "./islands/AddNewExpense.tsx": $$0,
